<feature>[front] Add accessToken refresh logic
- Modify middleware logic to reuse Authentication logic
- Create AuthLib.ts to reuse Authentication logic
- Modify Server-Action logic within authorization to refresh accessToken
